Slight code refactoring
